+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.2] - 2025-07-10
+
+### Changed
+
+- allowed SourceType 'AzureDatabricksDeltaLakeSink' to 'AzureDatabricksDeltaLakeSource' in 'New-CopyActivity'
+- allowed SinkType 'DelimitedTextSource' to 'DelimitedTextSink' in 'New-CopyActivity'
+
 ## [0.6.1] - 2025-07-10
 
 ### Fixed